Move worker into its own file
